+++ src/parallel_shooter/enemy_factory.go
@@ -35,7 +35,7 @@ func NewEnemy1Factory(g *Game) *Enemy1Factory {
 }
 
 func (ef *Enemy1Factory) NewObject(x,y int, p Phase, c Commander) *Character {
-	o := Object{game:ef.game, point: NewPoint(x, y), height:10, width:10, phase: p, images: &ef.imageSet}
+	o := Object{game:ef.game, point: NewPoint(float64(x), float64(y)), height:10, width:10, phase: p, images: &ef.imageSet}
 	ca := CharacterAttr{hp: 1, score: 0, value: 100, shotImages: &ef.shotImageSet}
 	return NewCharacter(o, ca, c)
 }
